wip: integration of project manager in ENGINE

+++ Laura/src/Project/Scene/SceneManager.cpp
@@ -6,33 +6,33 @@ namespace Laura
 	LR_GUID SceneManager::CreateScene(const std::string& name) {
 		auto scene = std::make_shared<Scene>(name);
 		m_Scenes[scene->GetGuid()] = scene;
-		LOG_ENGINE_INFO("CreateScene: created new scene \"{0}\" with GUID {1}", name, scene->GetGuid());
+		LOG_ENGINE_INFO("CreateScene: created new scene \"{0}\" with GUID {1}", name, (uint64_t)scene->GetGuid());
 		return scene->GetGuid();
 	}
 
 	void SceneManager::DeleteScene(LR_GUID guid) {
 		if (m_OpenSceneGuid == guid) {
 			m_OpenSceneGuid = LR_GUID::INVALID;
-			LOG_ENGINE_INFO("DeleteScene: closed open scene (GUID {0}) because it was deleted", guid);
+			LOG_ENGINE_INFO("DeleteScene: closed open scene (GUID {0}) because it was deleted", (uint64_t)guid);
 		}
 
 		size_t removed = m_Scenes.erase(guid);
 		if (removed == 0) {
-			LOG_ENGINE_WARN("DeleteScene: no scene found with GUID {0}; nothing deleted", guid);
+			LOG_ENGINE_WARN("DeleteScene: no scene found with GUID {0}; nothing deleted", (uint64_t)guid);
 		} else {
-			LOG_ENGINE_INFO("DeleteScene: successfully removed scene with GUID {0}", guid);
+			LOG_ENGINE_INFO("DeleteScene: successfully removed scene with GUID {0}", (uint64_t)guid);
 		}
 	}
 
 	bool SceneManager::SetOpenScene(LR_GUID guid) {
 		auto scene = find(guid);
 		if (!scene) {
-			LOG_ENGINE_WARN("SetOpenScene: cannot open scene; no scene registered with GUID {0}", guid);
+			LOG_ENGINE_WARN("SetOpenScene: cannot open scene; no scene registered with GUID {0}", (uint64_t)guid);
 			return false;
 		}
 
 		m_OpenSceneGuid = guid;
-		LOG_ENGINE_INFO("SetOpenScene: now tracking scene with GUID {0} as the active scene", guid);
+		LOG_ENGINE_INFO("SetOpenScene: now tracking scene with GUID {0} as the active scene", (uint64_t)guid);
 		return true;
 	}
 
@@ -50,7 +50,7 @@ namespace Laura
 			}
 
 			if (!find(guid)) {
-				LOG_ENGINE_INFO("SaveScenesToFolder: deleting orphaned scene file \"{0}\" (GUID {1})", scenepath.string(), guid);
+				LOG_ENGINE_INFO("SaveScenesToFolder: deleting orphaned scene file \"{0}\" (GUID {1})", scenepath.string(), (uint64_t)guid);
 				std::filesystem::remove(scenepath);
 			}
 		}
@@ -59,9 +59,9 @@ namespace Laura
 		for (const auto& [guid, scene] : m_Scenes) {
 			std::filesystem::path scenepath = ComposeScenepathFromGuid(folderpath, guid);
 			if (!SaveSceneFile(scenepath, scene)) {
-				LOG_ENGINE_WARN("SaveScenesToFolder: failed to serialize scene GUID {0} to \"{1}\"", guid, scenepath.string());
+				LOG_ENGINE_WARN("SaveScenesToFolder: failed to serialize scene GUID {0} to \"{1}\"", (uint64_t)guid, scenepath.string());
 			} else {
-				LOG_ENGINE_INFO("SaveScenesToFolder: successfully saved scene GUID {0} to \"{1}\"", guid, scenepath.string());
+				LOG_ENGINE_INFO("SaveScenesToFolder: successfully saved scene GUID {0} to \"{1}\"", (uint64_t)guid, scenepath.string());
 			}
 		}
 	}
@@ -75,7 +75,7 @@ namespace Laura
 			}
 			LR_GUID guid = scene->GetGuid();
 			m_Scenes[guid] = scene;
-			LOG_ENGINE_INFO("LoadScenesFromFolder: loaded scene \"{0}\" with GUID {1}", scenepath.string(), guid);
+			LOG_ENGINE_INFO("LoadScenesFromFolder: loaded scene \"{0}\" with GUID {1}", scenepath.string(), (uint64_t)guid);
 		}
 	}
 }
